requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
facebook-sdk
